Complete routes and tests for '/'
